adequacao das imagens

+++ Service/Interface/IReceitaService.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Net.Http;
 using System.Threading.Tasks;
@@ -10,7 +11,7 @@ namespace ReceitasDeSucesso.Service.Interface
         Task<IEnumerable<Receita>> ObterLista();
         Task<Receita> ObterItem(int? id);
         Task<HttpResponseMessage> InserirItem(Receita item);
-        Task<HttpResponseMessage> AlterarItem(int id, Receita item);
+        Task<HttpResponseMessage> AlterarItem(Int64 id, Receita item);
         Task<HttpResponseMessage> DeletarItem(int? id);
     }
 }
